fixed some error logging in reviews and likes

+++ store/likes/likeStore.js
@@ -1,6 +1,6 @@
 import { likeCollection } from "./likeSchema.js";
-import { recipeStore } from "../recipes/recipeStore.js";
 import { getDatabase } from "../database.js";
+import { ObjectId } from "mongodb";
 
 class LikeStore {
   constructor() {
@@ -10,7 +10,13 @@ class LikeStore {
   init() {
     const db = getDatabase();
     this.collection = db.collection(likeCollection);
+    this.recipeCollection = db.collection("recipes");
     this.recipeStatsCollection = db.collection("recipeStats");
+  }
+
+  async checkForRecipeId(recipeId) {
+    const id = new ObjectId(recipeId);
+    return Boolean(await this.recipeCollection.findOne({ _id: id }));
   }
 
   async checkForLike(userId, recipeId) {
@@ -18,41 +24,46 @@ class LikeStore {
   }
 
   async likeRecipe(userId, recipeId) {
-    try {
-      const alreadyLiked = await this.checkForLike(userId, recipeId);
-      if (alreadyLiked) {
-        throw new Error("Recipe has already been liked");
-      }
-
-      await this.collection.insertOne({
-        userId: userId,
-        recipeId: recipeId,
-        createdAt: new Date(),
-      });
-      await this.recipeStatsCollection.updateOne(
-        { recipeId },
-        { $inc: { likeCount: 1 } },
-        { upsert: true }
-      );
-    } catch (err) {
-      throw err;
+    const recipeCheck = await this.checkForRecipeId(recipeId);
+    if (!recipeCheck) {
+      throw new Error("RECIPE_NOT_FOUND");
     }
+
+    const alreadyLiked = await this.checkForLike(userId, recipeId);
+    if (alreadyLiked) {
+      throw new Error("Recipe has already been liked");
+    }
+
+    await this.collection.insertOne({
+      userId: userId,
+      recipeId: recipeId,
+      createdAt: new Date(),
+    });
+    await this.recipeStatsCollection.updateOne(
+      { recipeId },
+      { $inc: { likeCount: 1 } },
+      { upsert: true }
+    );
+    return { success: true };
   }
 
   async unlikeRecipe(userId, recipeId) {
-    try {
-      const alreadyLiked = await this.checkForLike(userId, recipeId);
-      if (!alreadyLiked) {
-        throw new Error("Cannot unlike a recipe that wasn't liked before");
-      }
-      await this.collection.deleteOne({ userId, recipeId });
-      await this.recipeStatsCollection.updateOne(
-        { recipeId },
-        { $inc: { likeCount: -1 } }
-      );
-    } catch (error) {
-      throw error;
+    const recipeCheck = await this.checkForRecipeId(recipeId);
+    if (!recipeCheck) {
+      throw new Error("RECIPE_NOT_FOUND");
     }
+
+    const alreadyLiked = await this.checkForLike(userId, recipeId);
+    if (!alreadyLiked) {
+      throw new Error("Cannot unlike a recipe that wasn't liked before");
+    }
+
+    await this.collection.deleteOne({ userId, recipeId });
+    await this.recipeStatsCollection.updateOne(
+      { recipeId },
+      { $inc: { likeCount: -1 } }
+    );
+    return { success: true };
   }
 }
